+++ lib/screens/Profile/View/room_details.dart
@@ -16,7 +16,7 @@ class RoomDetails extends StatefulWidget {
   final DateTime? checkIn;
   final DateTime? checkOut;
   final int quantity;
-  final int userPointsBalance;
+  final double userPointsBalance;
   final String ownerLocation;
   final String ownerUnitNo;
   final String bookingLocationName;
@@ -105,7 +105,7 @@ class _RoomDetailsState extends State<RoomDetails> {
     }
   }
 
-  int totalPoints() => widget.room.roomTypePoints;
+  double totalPoints() => widget.room.roomTypePoints;
 
   String formattedTotalPoints() {
     final formatter = NumberFormat('#,###');
